Change all `int`s by `ll`

+++ content/combinatorial/IntPerm.h
@@ -8,9 +8,9 @@
  */
 #pragma once
 
-int permToInt(vi& v) {
-	int use = 0, i = 0, r = 0;
-	for(int x:v) r = r * ++i + __builtin_popcount(use & -(1<<x)),
+ll permToInt(vi& v) {
+	ll use = 0, i = 0, r = 0;
+	for(ll x:v) r = r * ++i + __builtin_popcount(use & -(1<<x)),
 		use |= 1 << x;                     // (note: minus, not ~!)
 	return r;
 }
