+++ content/data-structures/FenwickTree2d.h
@@ -14,20 +14,20 @@
 
 struct FT2 {
 	vector<vi> ys; vector<FT> ft;
-	FT2(int limx) : ys(limx) {}
-	void fakeUpdate(int x, int y) {
+	FT2(ll limx) : ys(limx) {}
+	void fakeUpdate(ll x, ll y) {
 		for (; x < sz(ys); x |= x + 1) ys[x].push_back(y);
 	}
 	void init() {
 		for (vi& v : ys) sort(all(v)), ft.emplace_back(sz(v));
 	}
-	int ind(int x, int y) {
-		return (int)(lower_bound(all(ys[x]), y) - ys[x].begin()); }
-	void update(int x, int y, ll dif) {
+	ll ind(ll x, ll y) {
+		return (ll)(lower_bound(all(ys[x]), y) - ys[x].begin()); }
+	void update(ll x, ll y, ll dif) {
 		for (; x < sz(ys); x |= x + 1)
 			ft[x].update(ind(x, y), dif);
 	}
-	ll query(int x, int y) {
+	ll query(ll x, ll y) {
 		ll sum = 0;
 		for (; x; x &= x - 1)
 			sum += ft[x-1].query(ind(x-1, y));
