+++ content/data-structures/UnionFind.h
@@ -10,11 +10,11 @@
 
 struct UF {
 	vi e;
-	UF(int n) : e(n, -1) {}
-	bool sameSet(int a, int b) { return find(a) == find(b); }
-	int size(int x) { return -e[find(x)]; }
-	int find(int x) { return e[x] < 0 ? x : e[x] = find(e[x]); }
-	bool join(int a, int b) {
+	UF(ll n) : e(n, -1) {}
+	bool sameSet(ll a, ll b) { return find(a) == find(b); }
+	ll size(ll x) { return -e[find(x)]; }
+	ll find(ll x) { return e[x] < 0 ? x : e[x] = find(e[x]); }
+	bool join(ll a, ll b) {
 		a = find(a), b = find(b);
 		if (a == b) return false;
 		if (e[a] > e[b]) swap(a, b);
