+++ content/geometry/ConvexHull.h
@@ -25,8 +25,8 @@ vector<P> convexHull(vector<P> pts) {
 	if (sz(pts) <= 1) return pts;
 	sort(all(pts));
 	vector<P> h(sz(pts)+1);
-	int s = 0, t = 0;
-	for (int it = 2; it--; s = --t, reverse(all(pts)))
+	ll s = 0, t = 0;
+	for (ll it = 2; it--; s = --t, reverse(all(pts)))
 		for (P p : pts) {
 			while (t >= s + 2 && h[t-2].cross(h[t-1], p) <= 0) t--;
 			h[t++] = p;
